pyproject.toml
[project]
name = "policy-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pydrive>=1.3.1",
    "pymupdf>=1.25.5",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=4.1.0",
]
